+++ src/lib/pocketbase.ts
@@ -1,5 +1,0 @@
-import PocketBase from 'pocketbase';
-
-export const pb = new PocketBase('https://db.endurnet.ch');
-
-pb.autoCancellation(false);